the server wont save my progress
hi i have been playing your server a few days now and its rely fun but the other day i logged on and i went into a dungeon. im not very good at dungeons so it took me a long time to get out. when i left i folowd a trail and did 2 quests and went up two levels when i logged out and logged back on later that day i was back in the dungeon i was back at level 11 again and all the stuff i had collected after i left the dungeon was gone. now i was as you can imagin anoyed but decided to give the server the beniffit of the doubt and kept playing.i got myself back up to the stats i was at b4 and i logged off. when i came back on it had happend again. this is extremely anoying and i am not doing thos quests for a third time also i am stuck in the dungeon again and it takes me for ever to get out